+++ client/tx_config.go
@@ -44,6 +44,7 @@ type (
 		SetGasLimit(limit uint64)
 		SetTip(tip *tx.Tip)
 		SetTimeoutHeight(height uint64)
+		SetFeePayer(feePayer sdk.AccAddress)
 		SetFeeGranter(feeGranter sdk.AccAddress)
 	}
 )