+++ x/auth/middleware/testutil_test.go
@@ -26,8 +26,9 @@ import (
 
 // testAccount represents an account used in the tests in x/auth/middleware.
 type testAccount struct {
-	acc  authtypes.AccountI
-	priv cryptotypes.PrivKey
+	acc    authtypes.AccountI
+	priv   cryptotypes.PrivKey
+	accNum uint64
 }
 
 // MWTestSuite is a test suite to be used with middleware tests.
@@ -89,25 +90,23 @@ func (s *MWTestSuite) SetupTest(isCheckTx bool) sdk.Context {
 
 // createTestAccounts creates `numAccs` accounts, and return all relevant
 // information about them including their private keys.
-func (s *MWTestSuite) createTestAccounts(ctx sdk.Context, numAccs int) []testAccount {
+func (s *MWTestSuite) createTestAccounts(ctx sdk.Context, numAccs int, coins sdk.Coins) []testAccount {
 	var accounts []testAccount
 
 	for i := 0; i < numAccs; i++ {
 		priv, _, addr := testdata.KeyTestPubAddr()
 		acc := s.app.AccountKeeper.NewAccountWithAddress(ctx, addr)
-		err := acc.SetAccountNumber(uint64(i))
+		accNum := uint64(i)
+		err := acc.SetAccountNumber(accNum)
 		s.Require().NoError(err)
 		s.app.AccountKeeper.SetAccount(ctx, acc)
-		someCoins := sdk.Coins{
-			sdk.NewInt64Coin("atom", 10000000),
-		}
-		err = s.app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, someCoins)
+		err = s.app.BankKeeper.MintCoins(ctx, minttypes.ModuleName, coins)
 		s.Require().NoError(err)
 
-		err = s.app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr, someCoins)
+		err = s.app.BankKeeper.SendCoinsFromModuleToAccount(ctx, minttypes.ModuleName, addr, coins)
 		s.Require().NoError(err)
 
-		accounts = append(accounts, testAccount{acc, priv})
+		accounts = append(accounts, testAccount{acc, priv, accNum})
 	}
 
 	return accounts